requirements.txt
# Core Framework Dependencies
streamlit
neo4j
python-dotenv

# LLM and AI Dependencies
langchain-google-genai
langchain-core

# Data Processing Dependencies
PyPDF2
pandas
requests

# Visualization and Monitoring
plotly
watchdog